Fix: GPGPUProperties TotalGlobalMem changed to long (required for GPUs with 2GB or more).
git-svn-id: https://cudafy.svn.codeplex.com/svn@8236 2b2adeab-ffde-484a-ad65-8eada40eade2

+++ Cudafy/Cudafy.Host/GPGPUProperties.cs
@@ -211,6 +211,6 @@ namespace Cudafy.Host
         /// Gets the total global memory.
         /// </summary>
         /// <value>The total global memory.</value>
-        public int TotalGlobalMem { get; internal set; }
+        public long TotalGlobalMem { get; internal set; }
     }
 }
